# swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"IUser": {
			"properties": {
				"_id": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"password": {
					"type": "string"
				},
				"role": {
					"type": "string"
				}
			},
			"required": [
				"_id",
				"email",
				"password",
				"role"
			],
			"type": "object",
			"additionalProperties": true
		},
		"loginModel": {
			"properties": {
				"token": {
					"type": "string"
				},
				"user": {
					"$ref": "#/definitions/IUser"
				}
			},
			"required": [
				"token",
				"user"
			],
			"type": "object",
			"additionalProperties": true
		}
	},
	"info": {
		"title": "dev-challenge",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		}
	},
	"paths": {
		"/user": {
			"get": {
				"operationId": "GetAll",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"items": {
								"$ref": "#/definitions/IUser"
							},
							"type": "array"
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin",
							"manager",
							"user"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"jwt": [
							"admin",
							"manager"
						]
					}
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"email": {
									"type": "string"
								},
								"role": {
									"type": "string"
								}
							},
							"title": "CreateBody",
							"type": "object",
							"required": [
								"email",
								"role"
							]
						}
					}
				]
			}
		},
		"/user/{id}": {
			"put": {
				"operationId": "Update",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"jwt": [
							"admin",
							"manager"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"email": {
									"type": "string"
								},
								"role": {
									"type": "string"
								}
							},
							"title": "UpdateBody",
							"type": "object",
							"required": [
								"email",
								"role"
							]
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/login": {
			"post": {
				"operationId": "Login",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/loginModel"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"email": {
									"type": "string"
								},
								"password": {
									"type": "string"
								}
							},
							"title": "LoginBody",
							"type": "object",
							"required": [
								"email",
								"password"
							]
						}
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {
		"api_key": {
			"type": "apiKey",
			"name": "access_token",
			"in": "query"
		},
		"tsoa_auth": {
			"type": "oauth2",
			"authorizationUrl": "http://swagger.io/api/oauth/dialog",
			"flow": "implicit",
			"scopes": {
				"write:pets": "modify things",
				"read:pets": "read things"
			}
		}
	}
}